+++ packages/blocks-core/js/libs/wordpress/list-item/test/functionality.blocks-2.e2e.cy.js
@@ -33,6 +33,16 @@ describe('List Item Block → Functionality + Inner blocks', () => {
 
 		cy.checkBlockCardItems(['normal', 'hover', 'marker']);
 
+		cy.checkBlockStatesPickerItems([
+			'elements/link',
+			'elements/bold',
+			'elements/italic',
+			'elements/kbd',
+			'elements/code',
+			'elements/span',
+			'elements/mark',
+		]);
+
 		//
 		// 1. Edit Block
 		//